Remove task table association
