+++ sqle/model/audit_plan_report.go
@@ -3,11 +3,12 @@ package model
 type AuditPlanReportV2 struct {
 	Model
 	AuditPlanID uint `json:"audit_plan_id" gorm:"index"`
-	TaskID      uint `json:"task_id" gorm:"index"`
 
 	AuditPlan           *AuditPlan              `gorm:"foreignkey:AuditPlanID"`
 	AuditPlanReportSQLs []*AuditPlanReportSQLV2 `gorm:"foreignkey:AuditPlanReportID"`
-	Task                *Task                   `gorm:"foreignkey:TaskID"`
+	PassRate            float64                 `json:"pass_rate"`
+	Score               int32                   `json:"score"`
+	AuditLevel          string                  `json:"audit_level"`
 }
 
 func (a AuditPlanReportV2) TableName() string {
