+++ sqle/model/audit_plan_report_list.go
@@ -1,15 +1,17 @@
 package model
 
+import "database/sql"
+
 type AuditPlanReportListDetail struct {
-	ID         string  `json:"id"`
-	AuditLevel string  `json:"audit_level"`
-	Score      int32   `json:"score"`
-	PassRate   float64 `json:"pass_rate"`
-	CreateAt   string  `json:"created_at"`
+	ID         string          `json:"id"`
+	AuditLevel sql.NullString  `json:"audit_level"`
+	Score      sql.NullInt32   `json:"score"`
+	PassRate   sql.NullFloat64 `json:"pass_rate"`
+	CreateAt   string          `json:"created_at"`
 }
 
 var auditPlanReportQueryTpl = `
-SELECT reports.id, tasks.score , tasks.pass_rate, tasks.audit_level, reports.created_at
+SELECT reports.id, reports.score , reports.pass_rate, reports.audit_level, reports.created_at
 
 {{- template "body" . -}} 
 
@@ -29,12 +31,12 @@ var auditPlanReportBodyTpl = `
 
 FROM audit_plan_reports_v2 AS reports
 JOIN audit_plans ON audit_plans.id = reports.audit_plan_id
-LEFT JOIN tasks ON reports.task_id = tasks.id
 
 WHERE reports.deleted_at IS NULL
 AND audit_plans.deleted_at IS NULL
-AND tasks.deleted_at IS NULL
 AND audit_plans.name = :audit_plan_name
+
+ORDER BY reports.created_at DESC , reports.id DESC 
 
 {{ end }}
 `
