+++ sqle/server/auditplan/task.go
@@ -102,7 +102,12 @@ func (at *baseTask) audit(task *model.Task) (*model.AuditPlanReportV2, error) {
 		return nil, err
 	}
 
-	auditPlanReport := &model.AuditPlanReportV2{AuditPlanID: at.ap.ID, TaskID: task.ID, Task: task}
+	auditPlanReport := &model.AuditPlanReportV2{
+		AuditPlanID: at.ap.ID,
+		PassRate:    task.PassRate,
+		Score:       task.Score,
+		AuditLevel:  task.AuditLevel,
+	}
 	for _, executeSQL := range task.ExecuteSQLs {
 		auditPlanReport.AuditPlanReportSQLs = append(auditPlanReport.AuditPlanReportSQLs, &model.AuditPlanReportSQLV2{
 			SQL:         executeSQL.Content,
